update observation model
